Add types module with is_instance_of_typed_dict function
Add a new `types` module containing the `is_instance_of_typed_dict` function
to check if an object is an instance of a specified TypedDict class.

The function uses `get_type_hints` to retrieve the type hints of the TypedDict
class and compares the object's keys and value types with the type hints. It
also checks if the TypedDict is total and requires all keys to be present.